requirements.txt
# Core dependencies for FaceNet training
torch==2.7.1
torchvision==0.22.1

# Dataset download dependencies
requests==2.32.4
tqdm==4.67.1

# Training and evaluation dependencies
numpy==2.1.2
matplotlib==3.10.3
seaborn==0.13.2

# Additional dependencies for LFW benchmark
scikit-learn==1.7.1
pandas==2.2.3
Pillow==11.1.0

#MTCNN dependencies
mtcnn==1.0.0
tensorflow==2.19.0